Use React.StrictMode
